Term 2 Seminar 2 Codes

+++ term2/seminar02_encapsulation/classroom_tasks/code/2string/04operator_addition_solution2.cpp
@@ -3,7 +3,6 @@
 using std::cout, std::endl;
 
 
-
 class String 
 {
 private:
@@ -17,7 +16,7 @@ public:
 
     String(const char* str) 
     {
-        int i = 0;
+        size_t i = 0;
         while (str[i] != '\0')
             i++;
         mSize = i;
@@ -25,7 +24,7 @@ public:
 
         mpData = (char*)std::malloc(sizeof(char) * mCapacity);
 
-        for (int i = 0; str[i]; i++)
+        for (size_t i = 0; str[i]; i++)
             mpData[i] = str[i];
         mpData[mSize] = '\0';
     }
